Temporary fix to whitelist.
The whitelist names were supposed to be read from a txt file in the mod's resource folder, but creating that folder seems to make the game crash when loading SkyBot in Forge. Until this is fixed, the names for the whitelist will be those that are detected at the moment of activation of the whitelist mod.