Refactor: Modules now use class name as identifier
Instead of any name that was given to them. The previous way of naming was confusing, and had to be kept consistent across several locations. Now class names can change without having to look through the entire code for occurrences, or getting null pointer exceptions. Plus, there's no room for errors with typos now, as if you get the class name wrong it won't compile.